% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidiana.R
\name{Euclidiana}
\alias{Euclidiana}
\title{Calcula distancia Mahalabonis}
\usage{
Euclidiana(Reales, Simulados)
}
\arguments{
\item{Reales}{dataFrame datos Reles}

\item{Simulados}{dataFrame datos Simulados}
}
\value{
Distancia Euclidiana entre \code{Reales} and \code{Simulados}.
}
\description{
Calcula distancia Mahalabonis
}
