% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mahalanobis.R
\name{Mahalanobis}
\alias{Mahalanobis}
\title{Calcula distancia Mahalabonis}
\usage{
Mahalanobis(Reales, Simulados)
}
\arguments{
\item{Reales}{dataFrame datos Reles}

\item{Simulados}{dataFrame datos Simulados}
}
\value{
Distancia Mahalanobis entre \code{Reales} and \code{Simulados}.
}
\description{
Calcula distancia Mahalabonis
}
\examples{
mahalanobis(data.frame(1:4, 4:7), data.frame(5:8, 3:6))
}
