% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan.R
\name{Manhattan}
\alias{Manhattan}
\title{Calcula distancia Manhattan}
\usage{
Manhattan(Reales, Simulados)
}
\arguments{
\item{Reales}{dataFrame datos Reles}

\item{Simulados}{dataFrame datos Simulados}
}
\value{
Distancia Manhattan entre \code{Reales} and \code{Simulados}.
}
\description{
Calcula distancia Manhattan
}
